+++ tests/models/nn/transformers/test_sasrec.py
@@ -38,6 +38,7 @@ from rectools.models.nn.transformers.negative_sampler import CatalogUniformSampl
 from rectools.models.nn.transformers.sasrec import SASRecDataPreparator, SASRecTransformerLayers
 from rectools.models.nn.transformers.similarity import DistanceSimilarityModule
 from rectools.models.nn.transformers.torch_backbone import TransformerTorchBackbone
+from rectools.models.nn.transformers.utils import leave_one_out_mask
 from tests.models.data import DATASET
 from tests.models.utils import (
     assert_default_config_and_default_model_params_are_the_same,
@@ -45,7 +46,7 @@ from tests.models.utils import (
 )
 from tests.testing_utils import assert_id_map_equal, assert_interactions_set_equal
 
-from .utils import custom_trainer, leave_one_out_mask
+from .utils import custom_trainer
 
 
 class TestSASRecModel:
@@ -239,7 +240,7 @@ class TestSASRecModel:
                 pd.DataFrame(
                     {
                         Columns.User: [10, 10, 10, 30, 30, 30, 40, 40, 40],
-                        Columns.Item: [12, 13, 11, 11, 12, 14, 12, 14, 11],
+                        Columns.Item: [13, 12, 11, 11, 12, 14, 14, 12, 11],
                         Columns.Rank: [1, 2, 3, 1, 2, 3, 1, 2, 3],
                     }
                 ),
@@ -344,7 +345,7 @@ class TestSASRecModel:
                 pd.DataFrame(
                     {
                         Columns.User: [10, 10, 30, 30, 30, 40, 40, 40],
-                        Columns.Item: [17, 15, 13, 14, 17, 13, 14, 15],
+                        Columns.Item: [17, 15, 13, 17, 14, 13, 14, 15],
                         Columns.Rank: [1, 2, 1, 2, 3, 1, 2, 3],
                     }
                 ),
@@ -355,7 +356,7 @@ class TestSASRecModel:
                 pd.DataFrame(
                     {
                         Columns.User: [10, 10, 30, 30, 30, 40, 40, 40],
-                        Columns.Item: [17, 15, 13, 14, 17, 13, 14, 15],
+                        Columns.Item: [17, 15, 13, 17, 14, 13, 14, 15],
                         Columns.Rank: [1, 2, 1, 2, 3, 1, 2, 3],
                     }
                 ),
@@ -366,7 +367,7 @@ class TestSASRecModel:
                 pd.DataFrame(
                     {
                         Columns.User: [10, 10, 30, 30, 30, 40, 40, 40],
-                        Columns.Item: [17, 15, 13, 14, 17, 13, 14, 15],
+                        Columns.Item: [17, 15, 13, 17, 14, 13, 14, 15],
                         Columns.Rank: [1, 2, 1, 2, 3, 1, 2, 3],
                     }
                 ),
@@ -411,9 +412,9 @@ class TestSASRecModel:
         (
             pd.DataFrame(
                 {
-                    Columns.User: [10, 10, 10, 30, 30, 30, 40, 40, 40],
-                    Columns.Item: [13, 17, 11, 11, 13, 15, 17, 13, 11],
-                    Columns.Rank: [1, 2, 3, 1, 2, 3, 1, 2, 3],
+                    Columns.User: [30, 30, 30, 40, 40, 40],
+                    Columns.Item: [11, 13, 17, 17, 13, 11],
+                    Columns.Rank: [1, 2, 3, 1, 2, 3],
                 }
             ),
         ),
@@ -439,7 +440,7 @@ class TestSASRecModel:
             similarity_module_type=DistanceSimilarityModule,
         )
         model.fit(dataset=dataset)
-        users = np.array([10, 30, 40])
+        users = np.unique(expected[Columns.User])
         actual = model.recommend(users=users, dataset=dataset, k=3, filter_viewed=False)
         pd.testing.assert_frame_equal(actual.drop(columns=Columns.Score), expected)
         pd.testing.assert_frame_equal(
@@ -452,9 +453,9 @@ class TestSASRecModel:
         (
             pd.DataFrame(
                 {
-                    Columns.User: [10, 10, 10, 30, 30, 30, 40, 40, 40],
-                    Columns.Item: [13, 12, 11, 11, 12, 13, 13, 14, 12],
-                    Columns.Rank: [1, 2, 3, 1, 2, 3, 1, 2, 3],
+                    Columns.User: [30, 30, 30, 40, 40, 40],
+                    Columns.Item: [11, 13, 12, 13, 14, 12],
+                    Columns.Rank: [1, 2, 3, 1, 2, 3],
                 }
             ),
         ),
@@ -480,7 +481,7 @@ class TestSASRecModel:
             similarity_module_type=DistanceSimilarityModule,
         )
         model.fit(dataset=dataset_item_features)
-        users = np.array([10, 30, 40])
+        users = np.unique(expected[Columns.User])
         actual = model.recommend(users=users, dataset=dataset_item_features, k=3, filter_viewed=False)
         pd.testing.assert_frame_equal(actual.drop(columns=Columns.Score), expected)
         pd.testing.assert_frame_equal(
@@ -556,9 +557,9 @@ class TestSASRecModel:
                 None,
                 pd.DataFrame(
                     {
-                        Columns.TargetItem: [12, 12, 12, 14, 14, 14, 17, 17, 17],
-                        Columns.Item: [12, 13, 14, 14, 12, 15, 17, 13, 14],
-                        Columns.Rank: [1, 2, 3, 1, 2, 3, 1, 2, 3],
+                        Columns.TargetItem: [12, 12, 12, 17, 17, 17],
+                        Columns.Item: [12, 13, 14, 17, 13, 14],
+                        Columns.Rank: [1, 2, 3, 1, 2, 3],
                     }
                 ),
             ),
@@ -567,9 +568,9 @@ class TestSASRecModel:
                 None,
                 pd.DataFrame(
                     {
-                        Columns.TargetItem: [12, 12, 12, 14, 14, 14, 17, 17, 17],
-                        Columns.Item: [13, 14, 11, 12, 15, 17, 13, 14, 11],
-                        Columns.Rank: [1, 2, 3, 1, 2, 3, 1, 2, 3],
+                        Columns.TargetItem: [12, 12, 12, 17, 17, 17],
+                        Columns.Item: [13, 14, 11, 13, 14, 11],
+                        Columns.Rank: [1, 2, 3, 1, 2, 3],
                     }
                 ),
             ),
@@ -607,7 +608,7 @@ class TestSASRecModel:
             similarity_module_type=DistanceSimilarityModule,
         )
         model.fit(dataset=dataset)
-        target_items = np.array([12, 14, 17])
+        target_items = np.unique(expected[Columns.TargetItem])
         actual = model.recommend_to_items(
             target_items=target_items,
             dataset=dataset,
@@ -835,6 +836,28 @@ class TestSASRecDataPreparator:
         return Dataset.construct(interactions_df)
 
     @pytest.fixture
+    def dataset_timestamp_preproc(self) -> Dataset:
+        interactions_df = pd.DataFrame(
+            [
+                [10, 13, 1, "2021-11-30"],
+                [10, 11, 1, "2021-11-29"],
+                [10, 12, 1, "2021-11-29"],
+                [30, 11, 1, "2021-11-27"],
+                [30, 12, 2, "2021-11-26"],
+                [30, 15, 1, "2021-11-25"],
+                [40, 11, 1, "2021-11-25"],
+                [40, 17, 1, "2021-11-26"],
+                [50, 16, 1, "2021-11-25"],
+                [10, 14, 1, "2021-11-28"],
+                [10, 16, 1, "2021-11-27"],
+                [20, 13, 9, "2021-11-28"],
+                [10, 17, 1, "2021-11-30"],
+            ],
+            columns=Columns.Interactions,
+        )
+        return Dataset.construct(interactions_df)
+
+    @pytest.fixture
     def data_preparator(self) -> SASRecDataPreparator:
         return SASRecDataPreparator(session_max_len=3, batch_size=4, dataloader_num_workers=0)
 
@@ -859,6 +882,68 @@ class TestSASRecDataPreparator:
             n_negatives=2,
             get_val_mask_func=get_val_mask_func,
         )
+
+    @pytest.mark.parametrize(
+        "val_users, expected_batch_train, expected_batch_val",
+        (
+            (
+                [10, 30],
+                {
+                    "x": torch.tensor([[5, 2, 3], [0, 0, 1], [0, 0, 2]]),
+                    "y": torch.tensor([[2, 3, 6], [0, 0, 3], [0, 0, 4]]),
+                    "yw": torch.tensor([[1.0, 1.0, 1.0], [0.0, 0.0, 2.0], [0.0, 0.0, 1.0]]),
+                    "unix_ts": torch.tensor(
+                        [
+                            [1638057600, 1638144000, 1638144000, 1638230400],
+                            [1637798400, 1637798400, 1637798400, 1637884800],
+                            [1637798400, 1637798400, 1637798400, 1637884800],
+                        ]
+                    ),
+                },
+                {
+                    "x": torch.tensor([[0, 1, 3], [2, 3, 6]]),
+                    "y": torch.tensor([[2], [4]]),
+                    "yw": torch.tensor([[1.0], [1.0]]),
+                    "unix_ts": torch.tensor(
+                        [
+                            [1637884800, 1637884800, 1637884800, 1637971200],
+                            [1638144000, 1638144000, 1638230400, 1638230400],
+                        ]
+                    ),
+                },
+            ),
+        ),
+    )
+    def test_process_unix_ts_aware(
+        self,
+        dataset_timestamp_preproc: Dataset,
+        val_users: tp.List,
+        expected_batch_train: tp.Dict[str, torch.Tensor],
+        expected_batch_val: tp.Dict[str, torch.Tensor],
+    ) -> None:
+        get_val_mask_func_kwargs = {"val_users": val_users}
+        data_preparator = SASRecDataPreparator(
+            session_max_len=3,
+            batch_size=4,
+            dataloader_num_workers=0,
+            add_unix_ts=True,
+            get_val_mask_func=leave_one_out_mask,
+            get_val_mask_func_kwargs=get_val_mask_func_kwargs,
+        )
+        data_preparator.process_dataset_train(dataset_timestamp_preproc)
+        assert "unix_ts" in data_preparator.train_dataset.interactions.df
+        assert data_preparator.val_interactions is not None
+        assert "unix_ts" in data_preparator.val_interactions
+        dataloader_train = data_preparator.get_dataloader_train()
+        train_iterator = next(iter(dataloader_train))
+        for key, value in train_iterator.items():
+            assert torch.equal(value, expected_batch_train[key])
+        dataloader_val = data_preparator.get_dataloader_val()
+        assert dataloader_val is not None
+        val_iterator = next(iter(dataloader_val))
+        for key, value in val_iterator.items():
+            if key == "unix_ts":
+                assert torch.equal(value, expected_batch_val[key])
 
     @pytest.mark.parametrize(
         "expected_user_id_map, expected_item_id_map, expected_train_interactions, expected_val_interactions",
@@ -1061,14 +1146,13 @@ class TestSASRecModelConfiguration:
                 "data_preparator_type": "rectools.models.nn.transformers.sasrec.SASRecDataPreparator",
                 "lightning_module_type": "rectools.models.nn.transformers.lightning.TransformerLightningModule",
                 "negative_sampler_type": "rectools.models.nn.transformers.negative_sampler.CatalogUniformSampler",
-                "get_val_mask_func": "tests.models.nn.transformers.utils.leave_one_out_mask",
+                "get_val_mask_func": "rectools.models.nn.transformers.utils.leave_one_out_mask",
                 "similarity_module_type": "rectools.models.nn.transformers.similarity.DistanceSimilarityModule",
                 "backbone_type": "rectools.models.nn.transformers.torch_backbone.TransformerTorchBackbone",
             }
             expected.update(simple_types_params)
             if use_custom_trainer:
                 expected["get_trainer_func"] = "tests.models.nn.transformers.utils.custom_trainer"
-
         assert actual == expected
 
     @pytest.mark.parametrize("use_custom_trainer", (True, False))